Update responseUtils to ts (#102)
* move sendMenuMessage in js form

* update sendMenuMessage to ts

* move sendSatisfactionSurvey as js

* update sendSatisfactionSurvey to ts

* move sendVotingStats as js

* update sendVotingStats to ts

* move sendInterimUpdate as js

* update sendInterimUpdate to ts

* move sendInterimPrompt as js

* move sendInterimPrompt as js

* update sendInterimPrompt to ts

* move respondToInstance as js

* update respondToInstance to ts

* complete responseUtils.ts conversion

* update
